Suppress NarrowingCompoundAssignment warnings (#2548)
Removes NarrowingCompoundAssignment compilation warnings introduced by PR #2539.

The flags field is now an int instead of a byte.

+++ reactor-netty5-http/src/main/java/reactor/netty5/http/client/AbstractHttpClientMetricsHandler.java
@@ -64,7 +64,7 @@ abstract class AbstractHttpClientMetricsHandler extends ChannelHandlerAdapter {
 
 	final Function<String, String> uriTagValue;
 
-	byte flags;
+	int flags;
 
 	final static int REQUEST_SENT = 0x01;
